navbar componetizado, añadidio scroll secundario a los cards de los anuncios para hacer que el mapa esté siempre visible
