+++ src/front/js/component/notification.js
@@ -1,0 +1,250 @@
+import React, { useContext } from "react";
+import { Context } from "../store/appContext";
+
+export const Notification = () => {
+    const { store } = useContext(Context);
+
+    return (
+
+        store.isLoggedIn && (
+            <div className="flex justify-center items-center">
+              <button
+                button
+                id="dropdownHoverButton1"
+                data-dropdown-toggle="dropdownHover1"
+                data-dropdown-trigger="hover"
+                className="inline-flex items-center text-sm font-medium text-center text-indigo-500 hover:text-indigo-700 focus:outline-none dark:hover:text-white dark:text-gray-400"
+                type="button"
+              >
+                <svg
+                  className="w-10 h-10"
+                  aria-hidden="true"
+                  fill="currentColor"
+                  viewBox="0 0 20 20"
+                  xmlns="http://www.w3.org/2000/svg"
+                >
+                  <path d="M10 2a6 6 0 00-6 6v3.586l-.707.707A1 1 0 004 14h12a1 1 0 00.707-1.707L16 11.586V8a6 6 0 00-6-6zM10 18a3 3 0 01-3-3h6a3 3 0 01-3 3z"></path>
+                </svg>
+                <div className="relative flex">
+                  <div className="relative inline-flex w-3 h-3 bg-red-500 border-2 border-white rounded-full bottom-4 right-4 dark:border-gray-900"></div>
+                </div>
+              </button>
+
+              <div
+                id="dropdownHover1"
+                className="z-20 hidden w-full max-w-sm bg-white divide-y divide-gray-100 rounded-lg shadow dark:bg-gray-800 dark:divide-gray-700"
+                aria-labelledby="dropdownHoverButton1"
+              >
+                <div className="block px-4 py-2 font-medium text-center text-gray-700 rounded-t-lg bg-indigo-50 dark:bg-gray-800 dark:text-white">
+                  Notificaciones
+                </div>
+                <div className="divide-y divide-gray-100 dark:divide-gray-700">
+                  <a
+                    href="#"
+                    className="flex px-4 py-3 hover:bg-indigo-100 dark:hover:bg-gray-700"
+                  >
+                    <div className="flex-shrink-0">
+                      <img
+                        className="rounded-full w-11 h-11"
+                        src="/docs/images/people/profile-picture-1.jpg"
+                        alt="Jese image"
+                      />
+                      <div className="absolute flex items-center justify-center w-5 h-5 ml-6 -mt-5 bg-blue-600 border border-white rounded-full dark:border-gray-800">
+                        <svg
+                          className="w-3 h-3 text-white"
+                          aria-hidden="true"
+                          fill="currentColor"
+                          viewBox="0 0 20 20"
+                          xmlns="http://www.w3.org/2000/svg"
+                        >
+                          <path d="M8.707 7.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l2-2a1 1 0 00-1.414-1.414L11 7.586V3a1 1 0 10-2 0v4.586l-.293-.293z"></path>
+                          <path d="M3 5a2 2 0 012-2h1a1 1 0 010 2H5v7h2l1 2h4l1-2h2V5h-1a1 1 0 110-2h1a2 2 0 012 2v10a2 2 0 01-2 2H5a2 2 0 01-2-2V5z"></path>
+                        </svg>
+                      </div>
+                    </div>
+                    <div className="w-full pl-3">
+                      <div className="text-gray-500 text-sm mb-1.5 dark:text-gray-400">
+                        New message from{" "}
+                        <span className="font-semibold text-gray-900 dark:text-white">
+                          Jese Leos
+                        </span>
+                        : "Hey, what's up? All set for the presentation?"
+                      </div>
+                      <div className="text-xs text-blue-600 dark:text-blue-500">
+                        a few moments ago
+                      </div>
+                    </div>
+                  </a>
+                  <a
+                    href="#"
+                    className="flex px-4 py-3 hover:bg-indigo-100 dark:hover:bg-gray-700"
+                  >
+                    <div className="flex-shrink-0">
+                      <img
+                        className="rounded-full w-11 h-11"
+                        src="/docs/images/people/profile-picture-1.jpg"
+                        alt="Jese image"
+                      />
+                      <div className="absolute flex items-center justify-center w-5 h-5 ml-6 -mt-5 bg-blue-600 border border-white rounded-full dark:border-gray-800">
+                        <svg
+                          className="w-3 h-3 text-white"
+                          aria-hidden="true"
+                          fill="currentColor"
+                          viewBox="0 0 20 20"
+                          xmlns="http://www.w3.org/2000/svg"
+                        >
+                          <path d="M8.707 7.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l2-2a1 1 0 00-1.414-1.414L11 7.586V3a1 1 0 10-2 0v4.586l-.293-.293z"></path>
+                          <path d="M3 5a2 2 0 012-2h1a1 1 0 010 2H5v7h2l1 2h4l1-2h2V5h-1a1 1 0 110-2h1a2 2 0 012 2v10a2 2 0 01-2 2H5a2 2 0 01-2-2V5z"></path>
+                        </svg>
+                      </div>
+                    </div>
+                    <div className="w-full pl-3">
+                      <div className="text-gray-500 text-sm mb-1.5 dark:text-gray-400">
+                        New message from{" "}
+                        <span className="font-semibold text-gray-900 dark:text-white">
+                          Jese Leos
+                        </span>
+                        : "Hey, what's up? All set for the presentation?"
+                      </div>
+                      <div className="text-xs text-blue-600 dark:text-blue-500">
+                        a few moments ago
+                      </div>
+                    </div>
+                  </a>
+                  <a
+                    href="#"
+                    className="flex px-4 py-3 hover:bg-indigo-100 dark:hover:bg-gray-700"
+                  >
+                    <div className="flex-shrink-0">
+                      <img
+                        className="rounded-full w-11 h-11"
+                        src="/docs/images/people/profile-picture-1.jpg"
+                        alt="Jese image"
+                      />
+                      <div className="absolute flex items-center justify-center w-5 h-5 ml-6 -mt-5 bg-blue-600 border border-white rounded-full dark:border-gray-800">
+                        <svg
+                          className="w-3 h-3 text-white"
+                          aria-hidden="true"
+                          fill="currentColor"
+                          viewBox="0 0 20 20"
+                          xmlns="http://www.w3.org/2000/svg"
+                        >
+                          <path d="M8.707 7.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l2-2a1 1 0 00-1.414-1.414L11 7.586V3a1 1 0 10-2 0v4.586l-.293-.293z"></path>
+                          <path d="M3 5a2 2 0 012-2h1a1 1 0 010 2H5v7h2l1 2h4l1-2h2V5h-1a1 1 0 110-2h1a2 2 0 012 2v10a2 2 0 01-2 2H5a2 2 0 01-2-2V5z"></path>
+                        </svg>
+                      </div>
+                    </div>
+                    <div className="w-full pl-3">
+                      <div className="text-gray-500 text-sm mb-1.5 dark:text-gray-400">
+                        New message from{" "}
+                        <span className="font-semibold text-gray-900 dark:text-white">
+                          Jese Leos
+                        </span>
+                        : "Hey, what's up? All set for the presentation?"
+                      </div>
+                      <div className="text-xs text-blue-600 dark:text-blue-500">
+                        a few moments ago
+                      </div>
+                    </div>
+                  </a>
+                  <a
+                    href="#"
+                    className="flex px-4 py-3 hover:bg-indigo-100 dark:hover:bg-gray-700"
+                  >
+                    <div className="flex-shrink-0">
+                      <img
+                        className="rounded-full w-11 h-11"
+                        src="/docs/images/people/profile-picture-1.jpg"
+                        alt="Jese image"
+                      />
+                      <div className="absolute flex items-center justify-center w-5 h-5 ml-6 -mt-5 bg-blue-600 border border-white rounded-full dark:border-gray-800">
+                        <svg
+                          className="w-3 h-3 text-white"
+                          aria-hidden="true"
+                          fill="currentColor"
+                          viewBox="0 0 20 20"
+                          xmlns="http://www.w3.org/2000/svg"
+                        >
+                          <path d="M8.707 7.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l2-2a1 1 0 00-1.414-1.414L11 7.586V3a1 1 0 10-2 0v4.586l-.293-.293z"></path>
+                          <path d="M3 5a2 2 0 012-2h1a1 1 0 010 2H5v7h2l1 2h4l1-2h2V5h-1a1 1 0 110-2h1a2 2 0 012 2v10a2 2 0 01-2 2H5a2 2 0 01-2-2V5z"></path>
+                        </svg>
+                      </div>
+                    </div>
+                    <div className="w-full pl-3">
+                      <div className="text-gray-500 text-sm mb-1.5 dark:text-gray-400">
+                        New message from{" "}
+                        <span className="font-semibold text-gray-900 dark:text-white">
+                          Jese Leos
+                        </span>
+                        : "Hey, what's up? All set for the presentation?"
+                      </div>
+                      <div className="text-xs text-blue-600 dark:text-blue-500">
+                        a few moments ago
+                      </div>
+                    </div>
+                  </a>
+                  <a
+                    href="#"
+                    className="flex px-4 py-3 hover:bg-indigo-100 dark:hover:bg-gray-700"
+                  >
+                    <div className="flex-shrink-0">
+                      <img
+                        className="rounded-full w-11 h-11"
+                        src="/docs/images/people/profile-picture-1.jpg"
+                        alt="Jese image"
+                      />
+                      <div className="absolute flex items-center justify-center w-5 h-5 ml-6 -mt-5 bg-blue-600 border border-white rounded-full dark:border-gray-800">
+                        <svg
+                          className="w-3 h-3 text-white"
+                          aria-hidden="true"
+                          fill="currentColor"
+                          viewBox="0 0 20 20"
+                          xmlns="http://www.w3.org/2000/svg"
+                        >
+                          <path d="M8.707 7.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l2-2a1 1 0 00-1.414-1.414L11 7.586V3a1 1 0 10-2 0v4.586l-.293-.293z"></path>
+                          <path d="M3 5a2 2 0 012-2h1a1 1 0 010 2H5v7h2l1 2h4l1-2h2V5h-1a1 1 0 110-2h1a2 2 0 012 2v10a2 2 0 01-2 2H5a2 2 0 01-2-2V5z"></path>
+                        </svg>
+                      </div>
+                    </div>
+                    <div className="w-full pl-3">
+                      <div className="text-gray-500 text-sm mb-1.5 dark:text-gray-400">
+                        New message from{" "}
+                        <span className="font-semibold text-gray-900 dark:text-white">
+                          Jese Leos
+                        </span>
+                        : "Hey, what's up? All set for the presentation?"
+                      </div>
+                      <div className="text-xs text-blue-600 dark:text-blue-500">
+                        a few moments ago
+                      </div>
+                    </div>
+                  </a>
+                </div>
+                <a
+                  href="#"
+                  className="block py-2 text-sm font-medium text-center text-gray-900 rounded-b-lg bg-indigo-50 hover:bg-indigo-100 dark:bg-gray-800 dark:hover:bg-gray-700 dark:text-white"
+                >
+                  <div className="inline-flex items-center ">
+                    <svg
+                      className="w-4 h-4 mr-2 text-gray-500 dark:text-gray-400"
+                      aria-hidden="true"
+                      fill="currentColor"
+                      viewBox="0 0 20 20"
+                      xmlns="http://www.w3.org/2000/svg"
+                    >
+                      <path d="M10 12a2 2 0 100-4 2 2 0 000 4z"></path>
+                      <path
+                        fillRule="evenodd"
+                        d="M.458 10C1.732 5.943 5.522 3 10 3s8.268 2.943 9.542 7c-1.274 4.057-5.064 7-9.542 7S1.732 14.057.458 10zM14 10a4 4 0 11-8 0 4 4 0 018 0z"
+                        clipRule="evenodd"
+                      ></path>
+                    </svg>
+                    Ver todas{" "}
+                  </div>
+                </a>
+              </div>
+            </div>
+        )
+    )
+
+}
